v2 cleanup
